# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Home Vault Expense Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface LoginRequest {
  /** Email or phone number */
  identifier: string;
  password: string;
}

export interface RegisterRequest {
  name: string;
  email?: string;
  phone?: string;
  password: string;
}

export interface User {
  id: string;
  name: string;
  email?: string;
  phone?: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Vault {
  id: string;
  name: string;
  description?: string;
  icon?: string;
  color?: string;
  balance: number;
  isMain: boolean;
  parentId?: string;
  ownerId: string;
  createdAt: string;
  transactionCount: number;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  credit: "credit",
  debit: "debit",
} as const;

export interface Transaction {
  id: string;
  vaultId: string;
  type: TransactionType;
  amount: number;
  description: string;
  imageUrl?: string;
  createdById: string;
  createdByName: string;
  createdAt: string;
  balanceAfter: number;
}

export type VaultMemberRole =
  (typeof VaultMemberRole)[keyof typeof VaultMemberRole];

export const VaultMemberRole = {
  admin: "admin",
  member: "member",
} as const;

export interface VaultMember {
  id: string;
  userId: string;
  vaultId: string;
  name: string;
  email?: string;
  phone?: string;
  role: VaultMemberRole;
  joinedAt: string;
}

export type VaultDetail = Vault & {
  transactions: Transaction[];
  members: VaultMember[];
};

export interface CreateVaultRequest {
  name: string;
  description?: string;
  icon?: string;
  color?: string;
  initialBalance?: number;
  isMain?: boolean;
  parentId?: string;
}

export interface UpdateVaultRequest {
  name?: string;
  description?: string;
  icon?: string;
  color?: string;
}

export type CreateTransactionRequestType =
  (typeof CreateTransactionRequestType)[keyof typeof CreateTransactionRequestType];

export const CreateTransactionRequestType = {
  credit: "credit",
  debit: "debit",
} as const;

export interface CreateTransactionRequest {
  type: CreateTransactionRequestType;
  amount: number;
  description: string;
  imageUrl?: string;
}

export type AddMemberRequestRole =
  (typeof AddMemberRequestRole)[keyof typeof AddMemberRequestRole];

export const AddMemberRequestRole = {
  admin: "admin",
  member: "member",
} as const;

export interface AddMemberRequest {
  /** Email or phone number of user to add */
  identifier: string;
  role: AddMemberRequestRole;
}

export type GetTransactionsParams = {
  search?: string;
};