# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Home Vault Expense Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterRequest {
  name: string;
  email?: string;
  phone?: string;
  password: string;
}